Fix: Pass scanner version correctly, fix test content paths
We don't want to maintain a scanner version string our wrapper/action. We therefore need to make the SonarSource/sonarqube-scan-action conditional, based on whether a specific scanner version was requested as an input. We also need to ensure the path handling is correct for the two test projects. The scans are not currently finding the test content. The local action test adds very little, and has
been removed. Also, some of the test-python-project setup steps are not strictly
required and just add complexity to the testing.yaml workflow logic, and have
been removed.

Signed-off-by: Matthew Watkins <mwatkins@linuxfoundation.org>